requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
cssselect2==0.8.0
Django==5.2
django-cors-headers==4.7.0
django-filer==3.3.1
django-filter==25.1
django-jazzmin==3.0.1
django-polymorphic==3.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
easy-thumbnails==2.10
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lxml==5.4.0
Markdown==3.8
packaging==25.0
pillow==11.3.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
reportlab==4.4.4
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
webencodings==0.5.1
